Move sessions db logic into a new file
